+++ webdev/test/daemon/app_domain_test.dart
@@ -3,29 +3,12 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @Timeout(Duration(minutes: 2))
-import 'dart:async';
 import 'dart:convert';
 
 import 'package:test/test.dart';
-import 'package:test_process/test_process.dart';
 
 import '../test_utils.dart';
 import 'utils.dart';
-
-Future<String> _getAppId(TestProcess webdev) async {
-  var appId = '';
-  while (await webdev.stdout.hasNext) {
-    var line = await webdev.stdout.next;
-    if (line.startsWith('[{"event":"app.started"')) {
-      line = line.substring(1, line.length - 1);
-      var message = json.decode(line) as Map<String, dynamic>;
-      appId = message['params']['appId'] as String;
-      break;
-    }
-  }
-  assert(appId.isNotEmpty);
-  return appId;
-}
 
 void main() {
   String exampleDirectory;
@@ -63,7 +46,7 @@ void main() {
       test('.log', () async {
         var webdev =
             await runWebDev(['daemon'], workingDirectory: exampleDirectory);
-        var appId = await _getAppId(webdev);
+        var appId = await waitForAppId(webdev);
         // The example app does an initial print.
         await expectLater(
             webdev.stdout,
@@ -78,7 +61,7 @@ void main() {
       test('.callServiceExtension', () async {
         var webdev =
             await runWebDev(['daemon'], workingDirectory: exampleDirectory);
-        var appId = await _getAppId(webdev);
+        var appId = await waitForAppId(webdev);
         var extensionCall = '[{"method":"app.callServiceExtension","id":0,'
             '"params" : { "appId" : "$appId", "methodName" : "ext.print"}}]';
         webdev.stdin.add(utf8.encode('$extensionCall\n'));
@@ -94,7 +77,7 @@ void main() {
       test('.reload', () async {
         var webdev =
             await runWebDev(['daemon'], workingDirectory: exampleDirectory);
-        var appId = await _getAppId(webdev);
+        var appId = await waitForAppId(webdev);
         var extensionCall = '[{"method":"app.restart","id":0,'
             '"params" : { "appId" : "$appId", "fullRestart" : false}}]';
         webdev.stdin.add(utf8.encode('$extensionCall\n'));
@@ -110,7 +93,7 @@ void main() {
       test('.restart', () async {
         var webdev =
             await runWebDev(['daemon'], workingDirectory: exampleDirectory);
-        var appId = await _getAppId(webdev);
+        var appId = await waitForAppId(webdev);
         var extensionCall = '[{"method":"app.restart","id":0,'
             '"params" : { "appId" : "$appId", "fullRestart" : true}}]';
         webdev.stdin.add(utf8.encode('$extensionCall\n'));
@@ -130,7 +113,7 @@ void main() {
       test('.stop', () async {
         var webdev =
             await runWebDev(['daemon'], workingDirectory: exampleDirectory);
-        var appId = await _getAppId(webdev);
+        var appId = await waitForAppId(webdev);
         var stopCall = '[{"method":"app.stop","id":0,'
             '"params" : { "appId" : "$appId"}}]';
         webdev.stdin.add(utf8.encode('$stopCall\n'));
