+++ webdev/test/daemon/utils.dart
@@ -16,6 +16,21 @@ Future<void> exitWebdev(TestProcess webdev) async {
   await webdev.exitCode;
 }
 
+Future<String> waitForAppId(TestProcess webdev) async {
+  var appId = '';
+  while (await webdev.stdout.hasNext) {
+    var line = await webdev.stdout.next;
+    if (line.startsWith('[{"event":"app.started"')) {
+      line = line.substring(1, line.length - 1);
+      var message = json.decode(line) as Map<String, dynamic>;
+      appId = message['params']['appId'] as String;
+      break;
+    }
+  }
+  assert(appId.isNotEmpty);
+  return appId;
+}
+
 Future<String> prepareWorkspace() async {
   var exampleDirectory = p.absolute(p.join(p.current, '..', 'example'));
